+++ test/terraform/workspace/outputs.tf
@@ -1,4 +1,0 @@
-output "workspace" {
-  description = "The name of the workspace."
-  value       = "${var.workspace}"
-}